SSLCA: Fix certificate validation
We should favour "-trusted" over "-CAfile" because it will skip the system-wide
CAs and ensure that the certificate is relay validated against the specified
CA.

For validation against an intermediate certificate, only an additional
"-partial_chain" is required. With "-untrusted" we previously added an
unstrusted intermediate certificate only and validated the cert against default
system wide installed CAs.