Packages: Add "pyapt" source type
Pyapt is a new source that is using the apt python bindings to parse the
Packages files from debian repositories. Compared to the python implementation
it is faster and more robust. It will use the dependencies of the newest
version of a package from a specific source (because it can use the python
bindings to compare the version numbers).